refactor(act): extract dry-run loop to runRestoreDryRun helper
Three adapters were running identical 25-line dry-run loops —
iterate, optionally validate, optionally drop snapshots, return a
RestoreResult. Same pattern as the validator dedup earlier in this
branch: same contract, three copies, zero adapter-specific logic.

Extract to `runRestoreDryRun(source, opts)` in
`libs/act/src/restore-validate.ts`. Adapters early-return when
`opts.dry_run` is set:

```ts
async restore(source, opts = {}) {
  if (opts.dry_run) return runRestoreDryRun(source, opts);
  // ...adapter-specific live path...
}
```

Net change: -43 per adapter (~129 LOC removed), +71 for the
helper. Adapter restore methods now own only the live wipe-and-
rebuild mechanics that differ across backends — PG's
`TRUNCATE … RESTART IDENTITY` + transaction, SQLite's
`DELETE + sqlite_sequence` inside `transaction("write")`,
InMemory's index swap + rollback snapshot. Everything else lives
in the framework.

## Stability charter impact

**Charter surface touched: `libs/act/src/index.ts`** indirectly
(re-exports the new `runRestoreDryRun` symbol from
`restore-validate.ts`).

**Category: additive.** New public function added to
`@rotorsoft/act`. Same justification as the prior commits on this
branch — additive surface gated by STABILITY.md's minor-bump policy,
no charter renames or removals. Both `validateRestoreRow` and
`runRestoreDryRun` ship together on the first 1.x minor release
carrying the restore options work; no external consumer can have
locked in a previous shape.

## Test plan

- `pnpm typecheck` — clean
- `pnpm test` — 1859 / 1859 (no test deletions; same shape as before
  this commit, just exercising the shared helper)
- Coverage: 100% statements / 100% branches / 100% functions / 100% lines.
