+++ libs/act-sqlite/src/sqlite-store.ts
@@ -18,6 +18,7 @@ import type {
   StreamPosition,
   StreamStats,
 } from "@rotorsoft/act";
+import { runRestoreDryRun } from "@rotorsoft/act";
 
 /**
  * SQLite store configuration
@@ -1051,46 +1052,8 @@ export class SqliteStore implements Store {
     opts: RestoreOptions = {}
   ): Promise<RestoreResult> {
     const started = Date.now();
-    const {
-      drop_snapshots = false,
-      dry_run = false,
-      on_progress,
-      validate,
-    } = opts;
-    // Dry-run path — iterate without opening a write transaction.
-    // Adapters don't know validation policy; if the caller passed
-    // `validate`, we collect per-row blockers and surface them on the
-    // result. The store is never touched in this branch.
-    if (dry_run) {
-      const errors: Array<{ row: number; reason: string }> = [];
-      let kept = 0;
-      let droppedSnapshots = 0;
-      let rowIdx = 0;
-      for await (const row of source) {
-        rowIdx++;
-        if (on_progress) on_progress({ processed: rowIdx });
-        if (validate) {
-          for (const r of validate(row, rowIdx))
-            errors.push({ row: rowIdx, reason: r.reason });
-        }
-        if (drop_snapshots && row.name === "__snapshot__") {
-          droppedSnapshots++;
-          continue;
-        }
-        kept++;
-      }
-      return {
-        kept,
-        duration_ms: Date.now() - started,
-        dropped: {
-          closed_streams: 0,
-          snapshots: droppedSnapshots,
-          empty_streams: 0,
-        },
-        dry_run: true,
-        errors,
-      };
-    }
+    if (opts.dry_run) return runRestoreDryRun(source, opts);
+    const { drop_snapshots = false, on_progress } = opts;
     const tx = await this.client.transaction("write");
     try {
       await tx.execute("DELETE FROM events");
