+++ libs/act/src/adapters/in-memory-store.ts
@@ -9,6 +9,7 @@
  * @category Adapters
  */
 import { DEFAULT_LANE, SNAP_EVENT, TOMBSTONE_EVENT } from "../ports.js";
+import { runRestoreDryRun } from "../restore-validate.js";
 import { ConcurrencyError } from "../types/errors.js";
 import type {
   BlockedLease,
@@ -953,47 +954,8 @@ export class InMemoryStore implements Store {
   ): Promise<RestoreResult> {
     await sleep();
     const started = Date.now();
-    const {
-      drop_snapshots = false,
-      dry_run = false,
-      on_progress,
-      validate,
-    } = opts;
-    // Dry-run path — iterate without mutating. Adapters don't know
-    // validation policy; if the caller passed `validate`, we collect
-    // per-row blockers and surface them on the result. No
-    // snapshot/rollback bookkeeping needed because we never touch
-    // `this._*` indexes.
-    if (dry_run) {
-      const errors: Array<{ row: number; reason: string }> = [];
-      let kept = 0;
-      let droppedSnapshots = 0;
-      let rowIdx = 0;
-      for await (const row of source) {
-        rowIdx++;
-        if (on_progress) on_progress({ processed: rowIdx });
-        if (validate) {
-          for (const r of validate(row, rowIdx))
-            errors.push({ row: rowIdx, reason: r.reason });
-        }
-        if (drop_snapshots && row.name === SNAP_EVENT) {
-          droppedSnapshots++;
-          continue;
-        }
-        kept++;
-      }
-      return {
-        kept,
-        duration_ms: Date.now() - started,
-        dropped: {
-          closed_streams: 0,
-          snapshots: droppedSnapshots,
-          empty_streams: 0,
-        },
-        dry_run: true,
-        errors,
-      };
-    }
+    if (opts.dry_run) return runRestoreDryRun(source, opts);
+    const { drop_snapshots = false, on_progress } = opts;
     // Snapshot every index so we can roll back on throw.
     const prevEvents = this._events;
     const prevStreams = this._streams;
